CPR-566 package reorganisation
