+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/seeding/PopulateFromProbationIntTest.kt
@@ -10,9 +10,9 @@ import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity.Compani
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.ReferenceEntity
 import uk.gov.justice.digital.hmpps.personrecord.model.types.IdentifierType
 import uk.gov.justice.digital.hmpps.personrecord.model.types.SourceSystemType.DELIUS
+import uk.gov.justice.digital.hmpps.personrecord.seeding.responses.allProbationCasesResponse
+import uk.gov.justice.digital.hmpps.personrecord.seeding.responses.allProbationCasesSingleResponse
 import uk.gov.justice.digital.hmpps.personrecord.test.randomCrn
-import uk.gov.justice.digital.hmpps.personrecord.test.responses.allProbationCasesResponse
-import uk.gov.justice.digital.hmpps.personrecord.test.responses.allProbationCasesSingleResponse
 import java.time.LocalDate
 
 @ActiveProfiles("seeding")
